Arranging the code
