+++ ChatClient/ChatClient/ChatUtils.cs
@@ -1,7 +1,4 @@
 ﻿using BasicChatContents;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace ChatClient
 {
