+++ ChatClient/ChatClient/Chats/GroupChat.cs
@@ -6,7 +6,6 @@ namespace ChatClient.Chats
 {
     public class GroupChat : BasicChat
     {
-        private const string NO_OTHER_USERS_CONNECTED = "No other users connected";
         private ChatUtils _chatUtils;
         public GroupChat(Byte[] bytes, Client tcpClient, ConsoleDisplayer consoleDisplayer)
             : base(bytes, tcpClient, consoleDisplayer)
