+++ ChatServer/ChatServer/Server.cs
@@ -25,7 +25,6 @@ namespace ChatServer
         private List<IChatManager> _allChatManagers;
         private ClientHandler _clientHandler;
         private PrivateChatHandler _privateChatHandler;
-        private GeneralHandler _generalHandler;
         private GroupChatHandler _groupChatHandler;
 
         public Server()
@@ -38,9 +37,8 @@ namespace ChatServer
             _globalChatManager = new GlobalChatManager(_globalChatId, _chatFunctions);
             _allChats = new List<Chat>();
             _allChatManagers = new List<IChatManager>();
-            _generalHandler = new GeneralHandler();
             _privateChatHandler = new PrivateChatHandler(_clients, _usersInChats, _allChats, _allChatManagers,  _clientHandler);
-            _groupChatHandler = new GroupChatHandler(_clients, _usersInChats, _allChats, _allChatManagers, _generalHandler, _clientHandler);
+            _groupChatHandler = new GroupChatHandler(_clients, _usersInChats, _allChats, _allChatManagers, _clientHandler);
         }
 
         public void StartListening()
